# chore: Optimize rego policy input allocations