--- a/coderd/rbac/input.json
+++ b/coderd/rbac/input.json
@@ -12,7 +12,21 @@
   },
   "subject": {
     "id": "10d03e62-7703-4df5-a358-4f76577d4e2f",
-    "roles": [],
+    "roles": [
+      {
+        "name": "owner",
+        "display_name": "Owner",
+        "site": [
+          {
+            "negate": false,
+            "resource_type": "*",
+            "action": "*"
+          }
+        ],
+        "org": {},
+        "user": [],
+      }
+    ],
     "groups": ["b617a647-b5d0-4cbe-9e40-26f89710bf18"],
     "scope": {
       "name": "Scope_all",